+++ src/main/java/de/maxhenkel/easyvillagers/blocks/tileentity/render/TraderRenderer.java
@@ -7,19 +7,21 @@ import de.maxhenkel.easyvillagers.blocks.tileentity.TraderTileentity;
 import de.maxhenkel.easyvillagers.blocks.tileentity.TraderTileentityBase;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.Blocks;
+import net.minecraft.block.GrindstoneBlock;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.BlockRendererDispatcher;
 import net.minecraft.client.renderer.IRenderTypeBuffer;
 import net.minecraft.client.renderer.RenderTypeLookup;
 import net.minecraft.client.renderer.entity.VillagerRenderer;
 import net.minecraft.client.renderer.tileentity.TileEntityRendererDispatcher;
+import net.minecraft.state.properties.AttachFace;
 import net.minecraft.util.Direction;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.vector.Vector3f;
 import net.minecraftforge.client.model.data.EmptyModelData;
 import net.minecraftforge.registries.ForgeRegistries;
 
-import javax.annotation.Nullable;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -63,14 +65,15 @@ public class TraderRenderer extends VillagerRendererBase<TraderTileentity> {
             matrixStack.scale(0.45F, 0.45F, 0.45F);
             matrixStack.translate(0.5D / 0.45D - 0.5D, 0D, 0.5D / 0.45D - 0.5D);
 
-            BlockState workstation = trader.getWorkstation().getDefaultState();
+            BlockState workstation = getState(trader.getWorkstation());
             BlockRendererDispatcher dispatcher = minecraft.getBlockRendererDispatcher();
             int color = minecraft.getBlockColors().getColor(workstation, null, null, 0);
             dispatcher.getBlockModelRenderer().renderModel(matrixStack.getLast(), buffer.getBuffer(RenderTypeLookup.func_239221_b_(workstation)), workstation, dispatcher.getModelForState(workstation), RenderUtils.getRed(color), RenderUtils.getGreen(color), RenderUtils.getBlue(color), combinedLight, combinedOverlay, EmptyModelData.INSTANCE);
             BlockState topBlock = getTopBlock(workstation);
-            if (topBlock != null) {
+            if (!topBlock.isAir()) {
                 matrixStack.translate(0D, 1D, 0D);
-                dispatcher.getBlockModelRenderer().renderModel(matrixStack.getLast(), buffer.getBuffer(RenderTypeLookup.func_239221_b_(topBlock)), topBlock, dispatcher.getModelForState(topBlock), RenderUtils.getRed(color), RenderUtils.getGreen(color), RenderUtils.getBlue(color), combinedLight, combinedOverlay, EmptyModelData.INSTANCE);
+                int topColor = minecraft.getBlockColors().getColor(topBlock, null, null, 0);
+                dispatcher.getBlockModelRenderer().renderModel(matrixStack.getLast(), buffer.getBuffer(RenderTypeLookup.func_239221_b_(topBlock)), topBlock, dispatcher.getModelForState(topBlock), RenderUtils.getRed(topColor), RenderUtils.getGreen(topColor), RenderUtils.getBlue(topColor), combinedLight, combinedOverlay, EmptyModelData.INSTANCE);
             }
             matrixStack.pop();
         }
@@ -78,31 +81,52 @@ public class TraderRenderer extends VillagerRendererBase<TraderTileentity> {
         matrixStack.pop();
     }
 
+    private static final Map<Block, BlockState> BLOCK_CACHE = new HashMap<>();
+
+    public static BlockState getState(Block block) {
+        BlockState state = BLOCK_CACHE.get(block);
+        if (state == null) {
+            BlockState fittingState = getFittingState(block);
+            BLOCK_CACHE.put(block, fittingState);
+            return fittingState;
+        } else {
+            return state;
+        }
+    }
+
+    protected static BlockState getFittingState(Block block) {
+        if (block == Blocks.GRINDSTONE) {
+            return block.getDefaultState().with(GrindstoneBlock.FACE, AttachFace.FLOOR);
+        }
+        return block.getDefaultState();
+    }
+
     public static final Map<ResourceLocation, ResourceLocation> TOP_BLOCKS = new HashMap<>();
-    private static final Map<ResourceLocation, BlockState> TOP_BLOCK_CACHE = new HashMap<>();
+    private static final Map<Block, BlockState> TOP_BLOCK_CACHE = new HashMap<>();
 
     static {
         TOP_BLOCKS.put(new ResourceLocation("car", "gas_station"), new ResourceLocation("car", "gas_station_top"));
     }
 
-    @Nullable
     protected static BlockState getTopBlock(BlockState bottom) {
-        ResourceLocation registryName = bottom.getBlock().getRegistryName();
-
-        ResourceLocation resourceLocation = TOP_BLOCKS.get(registryName);
-        if (resourceLocation == null) {
-            return null;
-        }
-        BlockState cached = TOP_BLOCK_CACHE.get(registryName);
+        BlockState cached = TOP_BLOCK_CACHE.get(bottom.getBlock());
         if (cached != null) {
             return cached;
         }
+        ResourceLocation resourceLocation = TOP_BLOCKS.get(bottom.getBlock().getRegistryName());
+        if (resourceLocation == null) {
+            BlockState state = Blocks.AIR.getDefaultState();
+            TOP_BLOCK_CACHE.put(bottom.getBlock(), state);
+            return state;
+        }
         Block b = ForgeRegistries.BLOCKS.getValue(resourceLocation);
         if (b == null) {
-            return null;
+            BlockState state = Blocks.AIR.getDefaultState();
+            TOP_BLOCK_CACHE.put(bottom.getBlock(), state);
+            return state;
         }
         BlockState state = b.getDefaultState();
-        TOP_BLOCK_CACHE.put(registryName, state);
+        TOP_BLOCK_CACHE.put(bottom.getBlock(), state);
         return state;
     }